Extract shared streaming-fetch utility from FhirApi.sendRequest
Pulls the fetch/streaming/abort/partial-body-parsing mechanics out of
FhirApi.sendRequest into a pure, auth/origin-agnostic utility
(sendStreamingRequest in src/utils/streamingFetch.ts) so a future
ConnectionFhirApi can reuse the same mechanics against a different
trust boundary without duplicating them. sendRequest still performs
its origin-pinning check, buildHeaders call, and handleUnauthorized
call; behavior and return shape are unchanged.
